requirements.txt
absl-py==2.1.0
albumentations==1.3.1
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.2
captum==0.8.0
certifi==2022.12.7
charset-normalizer==2.1.1
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==4.4.2
dill==0.3.9
et-xmlfile==1.1.0
exceptiongroup==1.2.2
executing==2.1.0
filelock==3.9.0
flatbuffers==23.5.26
fonttools==4.46.0
fsspec==2024.10.0
gast==0.5.4
google-auth==2.26.2
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grad-cam==1.5.5
grpcio==1.60.0
h5py==3.10.0
hdbscan==0.8.39
huggingface-hub==0.31.2
idna==3.4
imageio==2.33.1
imageio-ffmpeg==0.4.9
ipython==8.29.0
jedi==0.19.2
Jinja2==3.1.2
joblib==1.3.2
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
llvmlite==0.43.0
Markdown==3.5.2
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.7
MiniSom==2.3.3
ml-dtypes==0.2.0
moviepy==1.0.3
mplcursors==0.6
mpmath==1.3.0
networkx==3.0
numba==0.60.0
numpy==1.24.1
oauthlib==3.2.2
opencv-python==4.10.0.84
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.4
parso==0.8.4
Pillow==9.3.0
proglog==0.1.10
prompt_toolkit==3.0.48
protobuf==4.23.4
psutil==5.9.6
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.18.0
pynndescent==0.5.13
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
qudida==0.0.4
regex==2024.11.6
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.5
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.4
seaborn==0.13.0
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.12.9
timm==1.0.11
tokenizers==0.21.1
torch==2.1.1+cu118
torchaudio==2.1.1+cu118
torchcam==0.4.0
torchsummary==1.5.1
torchvision==0.16.1+cu118
tqdm==4.66.1
traitlets==5.14.3
transformers==4.51.3
ttach==0.0.3
typing_extensions==4.12.2
tzdata==2023.3
ultralytics==8.3.34
ultralytics-thop==2.0.12
umap==0.1.1
umap-learn==0.5.7
urllib3==1.26.13
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
XlsxWriter==3.2.2
